changed rating dtype from decimal to int

+++ database/migrations/2024_12_18_081016_create_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
                     $table->id();
                     $table->unsignedBigInteger('user_id');
                     $table->unsignedBigInteger('bookID');
-                    $table->decimal('rating', 3, 2)->nullable();
+                    $table->integer('rating')->nullable();
                     $table->string('comment', 255)->nullable();
                     $table->timestamps();
